backend/requirements.txt
django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.1.0
djangorestframework-simplejwt==5.2.2
python-dotenv==1.0.0
openai==0.27.8
pymongo==4.3.3
djongo==1.3.6
cryptography==41.0.1
gitpython==3.1.31
